requirements.txt
fastapi
uvicorn
mysql-connector-python
email-validator
python-dotenv
